requirements.txt
des sollte per hand gehen


jupyter
pytorch
numpy
matplotlib
nibabel
scikit-image
torchio
monai
synapseclient
tqdm